refactor(pdf): extract buildFitCSS at the 300-line limit
The atomic-diagram rule pushed pdfPrintCss.ts to 316 lines and the file-size
gate failed. Split rather than baselined.

The seam is responsibility: pdfPrintCss keeps the page-break and layout rules,
pdfFitToPage takes the sizing arithmetic that bounds oversized figures to the
printable height. 230 / 105 lines.

Verified: typecheck clean, file-size gate green, 831 tests across 48 files.

Refs #1284
